Change in controller

+++ AutoMapperDemo.WebAPI/Repositories/IStudentRepository.cs
@@ -5,7 +5,7 @@ namespace AutoMapperDemo.WebAPI.Repositories
     public interface IStudentRepository
     {
         IEnumerable<Student> GetAllStudents();
-        Student GetStudentById(int id);
+        Student GetStudentById(long id);
         void AddStudent(Student student);
         void UpdateStudent(Student student);
         void DeleteStudent(Student student);
