+++ AutoMapperDemo.WebAPI/Service/IStudentService.cs
@@ -5,7 +5,7 @@ namespace AutoMapperDemo.WebAPI.Service
     public interface IStudentService
     {
         IEnumerable<Student> GetAllStudents();
-        Student GetStudentById(int id);
+        Student GetStudentById(long id);
         void AddStudent(Student student);
         void UpdateStudent(Student student);
         void DeleteStudent(Student student);
